[Android]adjust manager position.
