--- Android/common/base/src/main/java/io/agora/scene/base/manager/HealthActivityManager.java
@@ -1,4 +1,0 @@
-package io.agora.scene.base.manager;
-
-public class HealthActivityManager {
-}
