+++ Android/scenes/ktv/src/main/java/io/agora/scene/ktv/live/fragment/dialog/MVFragment.java
@@ -14,12 +14,12 @@ import java.util.List;
 import io.agora.scene.base.component.BaseRecyclerViewAdapter;
 import io.agora.scene.base.component.BaseViewBindingFragment;
 import io.agora.scene.base.component.OnItemClickListener;
-import io.agora.scene.base.manager.RoomManager;
 import io.agora.scene.ktv.R;
 import io.agora.scene.ktv.databinding.FragmentDialogMvBinding;
 import io.agora.scene.ktv.databinding.KtvItemMvBinding;
 import io.agora.scene.ktv.live.RoomLivingActivity;
 import io.agora.scene.ktv.live.holder.MVHolder;
+import io.agora.scene.ktv.manager.RoomManager;
 import io.agora.scene.widget.DividerDecoration;
 
 public class MVFragment extends BaseViewBindingFragment<FragmentDialogMvBinding> implements OnItemClickListener<Integer> {
