+++ Android/scenes/ktv/src/main/java/io/agora/scene/ktv/manager/ResourceManager.java
@@ -1,4 +1,4 @@
-package io.agora.scene.base.manager;
+package io.agora.scene.ktv.manager;
 
 import android.content.Context;
