+++ Android/scenes/ktv/src/main/java/io/agora/scene/ktv/manager/SingleMusicPlayer.java
@@ -1,4 +1,4 @@
-package io.agora.scene.base.manager;
+package io.agora.scene.ktv.manager;
 
 import android.content.Context;
 import android.text.TextUtils;
@@ -17,6 +17,7 @@ import io.agora.scene.base.R;
 import io.agora.scene.base.api.model.User;
 import io.agora.scene.base.bean.MemberMusicModel;
 import io.agora.scene.base.data.model.AgoraMember;
+import io.agora.scene.base.manager.UserManager;
 import io.agora.scene.base.utils.ToastUtils;
 import io.reactivex.SingleObserver;
 import io.reactivex.android.schedulers.AndroidSchedulers;
