+++ Android/scenes/ktv/src/main/java/io/agora/scene/ktv/manager/bean/RTCMessageBean.java
@@ -1,4 +1,4 @@
-package io.agora.scene.base.bean.room;
+package io.agora.scene.ktv.manager.bean;
 
 public class RTCMessageBean {
     public long total;
